Upgraded codemirror and removed angular-ui.
Codemirror was behind quite a few versions, and the use of angular-ui
was keeping the upgrade from happening, as that module disappeared from
angular long ago and simply doesn't work with the new codemirror
anymore. This should make upgrading angular much easier, as well, since
this was the major sticking point.